Move query-content & query-setup to separate files
