+++ packages/block-library/src/comments-query/edit/query-content.js
@@ -2,7 +2,6 @@
  * WordPress dependencies
  */
 import { useSelect, useDispatch } from '@wordpress/data';
-import { cloneBlock } from '@wordpress/blocks';
 import { useInstanceId } from '@wordpress/compose';
 import { useEffect } from '@wordpress/element';
 import {
@@ -12,7 +11,6 @@ import {
 	useSetting,
 	store as blockEditorStore,
 	__experimentalUseInnerBlocksProps as useInnerBlocksProps,
-	__experimentalBlockPatternSetup as BlockPatternSetup,
 } from '@wordpress/block-editor';
 import { SelectControl } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
@@ -20,29 +18,8 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import QueryToolbar from './toolbar';
+import QueryToolbar from '../toolbar';
 
-/**
- * Recurses over a list of blocks and returns the clientId
- * of the first found Post Comments Query Loop block.
- *
- * @param {WPBlock[]} blocks The list of blocks to look through.
- * @return {string=} The clientId.
- */
-export const getFirstQueryClientIdFromBlocks = ( blocks ) => {
-	const blocksQueue = [ ...blocks ];
-	while ( blocksQueue.length > 0 ) {
-		const block = blocksQueue.shift();
-		if ( block.name === 'core/comments-query' ) {
-			return block.clientId;
-		}
-		block.innerBlocks?.forEach( ( innerBlock ) => {
-			blocksQueue.push( innerBlock );
-		} );
-	}
-};
-
-// XXX: Update the stub comments-template
 const TEMPLATE = [ [ 'core/comments-template' ] ];
 
 export function QueryContent( { attributes, setAttributes } ) {
@@ -77,6 +54,7 @@ export function QueryContent( { attributes, setAttributes } ) {
 			setAttributes( { queryId: instanceId } );
 		}
 	}, [ queryId, instanceId ] );
+
 	const updateQuery = ( newQuery ) =>
 		setAttributes( { query: { ...query, ...newQuery } } );
 	const updateDisplayLayout = ( newDisplayLayout ) =>
@@ -112,59 +90,3 @@ export function QueryContent( { attributes, setAttributes } ) {
 		</>
 	);
 }
-
-const QueryPlaceholder = () => {
-	const blockProps = useBlockProps();
-
-	// XXX: Here will be a proper placeholder
-	return (
-		<div { ...blockProps }> QueryPlacholder: Actually implement this </div>
-	);
-};
-
-const QueryPatternSetup = ( props ) => {
-	const { clientId, name: blockName } = props;
-	const blockProps = useBlockProps();
-	const { replaceBlock, selectBlock } = useDispatch( blockEditorStore );
-	const onBlockPatternSelect = ( blocks ) => {
-		const clonedBlocks = blocks.map( ( block ) => cloneBlock( block ) );
-
-		const firstQueryClientId = getFirstQueryClientIdFromBlocks(
-			clonedBlocks
-		);
-
-		replaceBlock( clientId, clonedBlocks );
-		if ( firstQueryClientId ) {
-			selectBlock( firstQueryClientId );
-		}
-	};
-
-	// `startBlankComponent` is what to render when clicking `Start blank`
-	// or if no matched patterns are found.
-	return (
-		<div { ...blockProps }>
-			<BlockPatternSetup
-				blockName={ blockName }
-				clientId={ clientId }
-				startBlankComponent={
-					// XXX: create a proper query placeholder here
-					<QueryPlaceholder { ...props } />
-				}
-				onBlockPatternSelect={ onBlockPatternSelect }
-			/>
-		</div>
-	);
-};
-
-const CommentsQueryEdit = ( props ) => {
-	const { clientId } = props;
-	const hasInnerBlocks = useSelect(
-		( select ) =>
-			!! select( blockEditorStore ).getBlocks( clientId ).length,
-		[ clientId ]
-	);
-	const Component = hasInnerBlocks ? QueryContent : QueryPatternSetup;
-	return <Component { ...props } />;
-};
-
-export default CommentsQueryEdit;
